recipes-initscripts/cml-boot/bootloader-loopdev: Create bootloader loopdev
When booting i.MX boards from SD card, the bootloader sits at a fixed
offset at the beginning of the SD card outside a proper partition. To
directly access the bootloader for updates, we create a loopdevice
pointing to this particular region of the SD card.

For modularity, this script is placed in an init script executed by
busybox instead of of putting it in the cml-boot script.

Signed-off-by: Johannes Wiesböck <johannes.wiesboeck@aisec.fraunhofer.de>